failing tests for dependencySatisfies
