+++ packages/macros/tests/module-present-test.ts
@@ -1,12 +1,8 @@
 import 'qunit';
-import { transform as transform6, TransformOptions as Options6 } from 'babel-core';
-import { transform as transform7, TransformOptions as Options7 } from '@babel/core';
-import { babelPluginConfig } from '..';
-import { join } from 'path';
-
+import { allBabelVersions, runDefault } from './helpers';
 const { test } = QUnit;
 
-function createTests(transform: (code: string) => string) {
+allBabelVersions(function (transform: (code: string) => string) {
   QUnit.module(`modulePresent`, function() {
 
     test('module is present', function(assert) {
@@ -92,32 +88,4 @@ function createTests(transform: (code: string) => string) {
       }, /the argument to modulePresent must be a string literal/);
     });
   });
-}
-
-const options7: Options7 = {
-  filename: join(__dirname, 'sample.js'),
-  presets: [],
-  plugins: [babelPluginConfig()]
-};
-
-const options6: Options6 = {
-  filename: join(__dirname, 'sample.js'),
-  presets: [],
-  plugins: [babelPluginConfig()]
-};
-
-QUnit.module('babel6', function() {
-  createTests(function(code: string){ return transform6(code, options6).code!; });
 });
-QUnit.module('babel7', function() {
-  createTests(function(code: string){ return transform7(code, options7)!.code!; });
-});
-
-function runDefault(code: string): any {
-  let cjsCode = transform7(code, {
-    plugins: ['@babel/plugin-transform-modules-commonjs']
-  })!.code!;
-  let exports = {};
-  eval(cjsCode);
-  return (exports as any).default();
-}
